refactor: removing braid atomic refs for locking in main thread
