+++ node/src/db/mod.rs
@@ -1,4 +1,6 @@
 #![allow(non_snake_case)]
+use std::collections::HashMap;
+
 use crate::bead::Bead;
 pub mod db_handlers;
 pub mod init_db;
@@ -8,10 +10,8 @@ pub mod init_db;
 pub enum InsertTupleTypes {
     InsertBeadSequentially {
         bead_to_insert: Bead,
-        txs_json: String,
-        relative_json: String,
-        parent_timestamp_json: String,
-        bead_id: usize,
+        curr_beads: Vec<Bead>,
+        bead_index_mapping: HashMap<bitcoin::BlockHash, usize>,
     },
 }
 #[derive(Debug, Clone)]
